refactor: move components like assets, editor, preview to frontend feature dir
